requirements.txt
asgiref==3.7.2
black==23.9.1
boto3==1.29.4
botocore==1.32.4
click==8.1.7
colorama==0.4.6
cssbeautifier==1.14.9
Django==4.2.5
django-filter==23.5
django-storages==1.14.2
django-widget-tweaks==1.5.0
EditorConfig==0.12.3
gunicorn==21.2.0
html-tag-names==0.1.2
html-void-elements==0.1.0
jmespath==1.0.1
jsbeautifier==1.14.9
json5==0.9.14
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.1
platformdirs==3.10.0
psycopg2==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.7
whitenoise==6.5.0
